Separate section + title accordion onto its own component that controls autosave
